requirements.txt
pytelegrambotapi==4.29.1
google-genai==1.39.1
python-dotenv==1.1.1
openai==1.35.1
transformers==4.57.0
fastmcp==2.12.4